+++ app/src/main/java/org/positive/daymotion/presentation/root/tabs/category/CategoryTabViewModel.kt
@@ -1,8 +1,8 @@
-package org.positive.daymotion.presentation.my
+package org.positive.daymotion.presentation.root.tabs.category
 
 import dagger.hilt.android.lifecycle.HiltViewModel
 import org.positive.daymotion.presentation.base.BaseViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class MyTabViewModel @Inject constructor() : BaseViewModel()
+class CategoryTabViewModel @Inject constructor() : BaseViewModel()
