+++ app/src/main/java/org/positive/daymotion/presentation/root/tabs/setting/SettingTabViewModel.kt
@@ -1,8 +1,8 @@
-package org.positive.daymotion.presentation.category
+package org.positive.daymotion.presentation.root.tabs.setting
 
 import dagger.hilt.android.lifecycle.HiltViewModel
 import org.positive.daymotion.presentation.base.BaseViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class CategoryTabViewModel @Inject constructor() : BaseViewModel()
+class SettingTabViewModel @Inject constructor() : BaseViewModel()
